requirements.txt
Jinja2==2.10.1
argparse==1.4.0
humanize>=0.5.1
libmozdata>=0.1.56
pep8==1.7.1
pyflakes==2.1.1
python-dateutil>=2.5.2
requests>=2.20.0
python-Levenshtein>=0.12.0
pytz>=2018.9
whatthepatch>=0.0.5
icalevents==0.1.20
icalendar==4.0.3
sqlalchemy==1.3.7
filelock==3.0.12
alembic==1.0.11
numpy==1.17.0
